refactor: move store logic into store module
